Fuse `test_atomic_wait.hpp` into `P1135R6_atomic_wait`.
No changes other than dropping the duplicate banner and `#pragma once`.
